DRTII-60 Remove old CrunchTestLike
